+++ packages/runtime/adapters/eventstore-adapters/eventstore-postgresql/src/connect.ts
@@ -6,9 +6,8 @@ import type {
   AdapterPoolPrimal,
   PostgresqlAdapterConfig,
 } from './types'
-import { ConnectionError, ServiceBusyError } from '@resolve-js/eventstore-base'
-import { isServiceBusyError } from './errors'
 import makePostgresClient from './make-postgres-client'
+import makeKnownError from './make-known-error'
 
 const connect = async (
   pool: AdapterPoolPrimal,
@@ -70,22 +69,8 @@ const connect = async (
       escape,
       connection,
     })
-
-    if (pool.executeStatement != null) {
-      await pool.executeStatement('SELECT 0 AS "defunct"')
-    }
   } catch (error) {
-    if (isServiceBusyError(error)) {
-      const busyError = new ServiceBusyError(error.message)
-      busyError.stack = error.stack
-      throw busyError
-    } else if (error instanceof Error) {
-      const connectionError = new ConnectionError(error.message)
-      connectionError.stack = error.stack
-      throw connectionError
-    } else {
-      throw error
-    }
+    throw makeKnownError(error)
   }
   log.debug('connection to postgres databases established')
 }