+++ packages/runtime/runtime/src/__mocks__/uuid/v4.js
@@ -1,2 +1,0 @@
-const mockedUUID = () => 'UUID'
-export default mockedUUID