+++ packages/runtime/runtime/src/common/defaults/json-serialize-state.js
@@ -1,9 +1,0 @@
-import { IS_BUILT_IN } from '@resolve-js/core'
-
-function serializeState(state) {
-  return JSON.stringify(state, null, 2)
-}
-
-serializeState[IS_BUILT_IN] = true
-
-export default serializeState