+++ packages/runtime/runtimes/runtime-aws-serverless/src/__mocks__/aws-sdk/clients/lambda.ts
@@ -1,0 +1,8 @@
+const Lambda = jest.fn()
+
+const promised = () =>
+  jest.fn().mockReturnValue({ promise: () => Promise.resolve() })
+
+Lambda.prototype.invoke = promised()
+
+export default Lambda