+++ packages/runtime/runtimes/runtime-base/src/handlers/index.ts
@@ -1,0 +1,7 @@
+export { commandHandler } from './command-handler'
+export { failHandler } from './fail-handler'
+export { liveRequireHandler } from './live-require-handler'
+export { mainHandler } from './main-handler'
+export { markupHandler } from './markup-handler'
+export { queryHandler } from './query-handler'
+export { uploaderHandler } from './uploader-handler'