+++ templates/js/vanilla/common/api-handlers/markup.js
@@ -1,5 +1,5 @@
-import getStaticBasedPath from '@resolve-js/runtime/lib/common/utils/get-static-based-path'
-import jsonUtfStringify from '@resolve-js/runtime/lib/common/utils/json-utf-stringify'
+import { getStaticBasedPath } from '@resolve-js/core'
+import { jsonUtfStringify } from '@resolve-js/core'
 const markupHandler = async (req, res) => {
   const { rootPath, staticPath, seedClientEnvs } = req.resolve
   const scriptUrl = getStaticBasedPath(rootPath, staticPath, 'index.js')