+++ CHANGELOG.md
@@ -2,7 +2,26 @@
 
 ## [Upcoming](https://github.com/reimagined/resolve/tree/HEAD)
 
-[Full Changelog](https://github.com/reimagined/resolve/compare/V0.31.10...HEAD)
+[Full Changelog](https://github.com/reimagined/resolve/compare/V0.32.0...HEAD)
+
+**New features:**
+
+- Add event loader in eventstore adapter inferface and loading events via pg-cursor [\#2067](https://github.com/reimagined/resolve/pull/2067)
+
+**Breaking changes:**
+
+- Refactored runtime libraries [\#2060](https://github.com/reimagined/resolve/pull/2060)
+
+**Other changes:**
+
+- Fix subscriber notification [\#2062](https://github.com/reimagined/resolve/pull/2062)
+- Migrate from "incremental import" to "import" [\#2061](https://github.com/reimagined/resolve/pull/2061)
+- Review strings in templates and examples [\#2059](https://github.com/reimagined/resolve/pull/2059)
+- Cleanup eventstore tests [\#2058](https://github.com/reimagined/resolve/pull/2058)
+
+## [V0.32.0](https://github.com/reimagined/resolve/tree/V0.32.0) (2021-10-01)
+
+[Full Changelog](https://github.com/reimagined/resolve/compare/V0.31.10...V0.32.0)
 
 **New features:**
 
@@ -25,6 +44,12 @@
 
 - Describe the `host` option [\#2043](https://github.com/reimagined/resolve/pull/2043)
 - Describe API for fundamental entities of reSolve [\#2032](https://github.com/reimagined/resolve/pull/2032)
+
+**Other changes:**
+
+- Next step on long builds [\#2056](https://github.com/reimagined/resolve/pull/2056)
+- Bump resolve-cloud-common version [\#2054](https://github.com/reimagined/resolve/pull/2054)
+- Update invocation type for send notifications [\#2053](https://github.com/reimagined/resolve/pull/2053)
 
 ## [V0.31.10](https://github.com/reimagined/resolve/tree/V0.31.10) (2021-09-08)